First-pass trajectory-accepting gRPC schema for platform
Draft a Protobuf schema for the platform to use in accepting trajectories
from mouse personalities. Note that I imagine using some other transport
(TBD) to carry Protobuf messages to the microcontroller -- no guarantees
I'll use Protobuf wire format either -- but wrapping it in a gRPC endpoint
on the Pi so that it does actually serve this RPC interface.