First-pass kinematics
Auto-generate a kinematics lookup table using Jupyter.

Note that I've plotted the six variables in heatmaps and they
all look pretty good -- at least in the DataFrames. I'm not exactly
sure that I've got the mapping to C++ struct array correct and that
I didn't, I dunno, invert an axis there?

Now to try integrating it into an Arduino sketch. That's gonna be a
whole thing :)