Get the 4260BL motor spinning; check in some other Arduino bringup code
I'm working on spinning the E-S Motor 4260BL. The 4260bl_pwm_encoder script
does a ramp of PWM and measures RPM in order to generate the linearity plot
in the Google Drive folder. I'm also checking in a few other bits of code
I've been using to bring up the Arduino, just in case they're useful.