Set microsteps on the driver
Zoda and I were observing that the motors were spinning surprisingly
slowly today; I'm confused about this because MS1/MS2 are low on both
drivers and so they should be using the default 1/8 microstep setting
already. So maybe the TMCStepper driver object is overwriting that
when we call begin()? I'm not sure.

Anyway, when we do explicitly set the microsteps to 1/8
using driver.microsteps(8), it spins at expected speed and so clearly
it is worth doing.