Get string length tracking working, index into LUT
The LUT is linking!!! Woo!!! File size is way up. All it took was
actually accessing the LUT and using it for something (so that the
compiler didn't optimize it out). This rocks.

Also encoding appears to be working too :) I'm seeing the right
number of pulse counts per revolution.

Next up: actually use the partial derivatives to track motion.

Note that I'm not sure this will give us the positional accuracy
we need -- we'll want to interpolate on X and Y in order to get
finer grain position so that we can properly orient to get into
mouse houses.