Update Poetry lockfile to reflect version 3.12.2
pyproject.toml has version 3.12.2 but the lockfile shows 3.12.3.
Re-lock to reflect so that Poetry will stop complaining.

Note that as I understand it, it's common practice to check in
lockfiles whenever they change; that way fellow developers and CI/CD
machines can stay in sync with you with requirements, target the
same versions, and check that your versions are available on other
platforms right away.