Implement a fun little squiggly G-Code path generator
Accept Zoda's paths in units of map grid spacing and poop out
a squiggly arc move that squiggles around in fun little 45-degree
arcs. I think it'll keep speed through corners decently well --
it squiggles to make sure that the mouse corners in an arc through
each 90 degree corner.

We'll probably want to throw this away later, and we'll need to be
careful about how we approach mouse houses (I think? we'll have to
unwind the arc into the house entrance?) but this should be a decent
start for moving around, and show the power of the platform.