Add pyenv .python-version for pi_controller
Tell pyenv to use 3.12.2 in the pi_controller repo to make setup
a little easier. Note that I'm putting this in the pi_controller
folder to be near the poetry.lock... we might want to promote that
up a level though as we might run things offboard on a laptop that
will also want to share the same requirements + Python version?