Update map.csv to 50 mm grid, Open Sauce configuration
The last map.csv was a best-guess draft; now we've actually
built the play structure for Open Sauce.

Also I'm down-rezzing to a 50 mm grid, which I think will
make navigation easier, give a little more room for error (not
try to steer the mouse close to obstacles)  and gives us a nice arc radius
for squigglin' :)

Note that for production we'll redesign the entire front panel
and customize the map per unit. This is just hard coding for Open
Sauce!!