Add super basic mouse house routines (UNTESTED)
For testing tomorrow, a few routines to enter and exit a mouse house.
These use the MotionCommand base class I put together and the basic
DirectMove and RelativeMove types.

This isn't tested!!! The mouse houses are still drying from being
glued a few hours ago. However I have a decent-looking CAD sketch done,
from which I got the measurements. So I think it has half a chance
of working as is... anyway we'll see tomorrow :)