Add Point-to-Node conversion
I'm going to stream live location from Moonraker in the hope of
getting the same update rate that Mainsail features, which is pretty
good. I think that just converting from the current reported location
to the nearest tile and saying "I'm here now" is a good way of figuring
out how far in the gcode queue we've gotten.

Toward that end, add a to_node() function to custommap which converts
back into map tile location (note that I'm starting to prefer the term
"tile" since it implies area! ie it's more intuitive to say "where am
I on the tile?" than it is to say "where am I in the node?").

Include the map size in MapConfig so that I can error-check in the
conversion function (just in case... I dunno. Just in case)