Add end_coordinates() to GCode objects
I'm trying to figure out where the motion stage is on the move sequence
it's been sent. I think the easiest way to do that is to take the current
position reported by the toolhead and back out which tile we're on.
We can then drop everything in the path that comes before that tile
(assuming the path doesn't cross itself, which it shouldn't I think).

Note that if the path crosses itself this will cause Weird Behavior!

Anyway I'm adding a method to GCode objects that gives us where the
GCode ends. I guess we could've inspected for an `x_mm` and `y_mm`
attribute but that sounds ugly and like it doesn't age well...