Store latest location from Moonraker
Now that we're subscribed to motion_output, crack the latest results
into a Point and save it behind a Lock.

This is gonna be a fairly reliable code path, I think, as it's used
exactly as-is to power the Mainsail frontend.

My notes on how I figured this out here:

https://feline.atlassian.net/wiki/spaces/FES/pages/76578818/Getting+Real-Time+Status+from+Klipper

I still need to figure out how to get this back to Zoda; I've added
a to_node method but haven't yet figured out where the right place
to invoke it is. That's a tomorrow problem I think~