Refactor "Node" to "Tile"
The map is a grid, yes? and we've been conceptualizing it as
an undirected graph of nodes (points) connected by regular
edges (four per internal node -- left, right, up, down).

However as we start to glue together the map and the motion system
this is feeling weird, like I'm associating physical points on the
board with the nearest node, and what I really want is to have the
map be of a series of squares each of which has actual area.

So I've found myself translating "node" to "tile" in my comments
and documentation and brain, which feels a lot more intuitive.

"Is this point contained within this tile?" --> yes! so intuitive.
Easy to reason about. "Is this point associated to this node?" is
just not so.

So hoping to rip off the bandaid now while we're fast and light,
do one fell swoop refactor and move on with our lives :)