Add Path object as interface between planning and machine
Take a crack at adding a nice ergonomic interface between path
planning and Klipper. This Path object is meant as a kind of drop-in
replacement for the Paths we've been generating, but adds methods
that hopefully let it be usefully shared and manipulated on both
ends.

The Klipper socket interface should call commit() when it submits
movements to the socket and advance() when it receives feedback
that the machine has advanced; the path planning should call
clear() / truncate_at() when it wants to re-plan and extend()
when it wants to add movement.

Still to figure out:
* How do we incorporate picking up and dropping off mice?
* Generate GCode / reconcile with SimpleSquigglePath
* Accept motion styles, speed, and other cues from path planning