Add position subscription and PathFollower object
Add a callback mechanism to MotionDriver and implement in MoonrakerSocket
so that we can get live updates as callbacks. Note that we will have to be
careful about threading with this callback system / not block in callbacks
or else we'll slow down the websocket connection!

Anyway, this lets me construct a PathFollower object which takes a
MotionDriver and executes Path following on it quite simply. (I thought
about making those utility functions on MotionDriver, but IDK, I figured
keeping MotionDriver as an abstract interface was cleaner...)

Subscription is currently not implemented on MockMotionDriver (I will do
this shortly / in an upcoming commit) or KlipperSocket
(which I might deprecate honestly as the websocket seems just fine).