Finish PathFollower by actually submitting motion; allow restart
This is turning into a cute little multi-threading project, with two
major threads interacting back and forth through the Path and
PathFollower object, namely the Klipper websocket receive thread
bringing in position updates and the path planning thread redoing
the upcoming path based on camera updates.

Add a MotionStyle ABC that allows swappable motion styles for going
from one tile to another. This actually has a serious design flaw
right now which is that it doesn't allow lookahead; SimpleSquiggleArcPath
does lookahead, which is really important and useful in order to
corner quickly, and I think we'll want to continue doing that. (We
can add this in some kludgy way, I just haven't yet... waiting for a
brain zap about what the "right" way to do it is.)

Add code to PathFollower to submit the first three tiles (three so
we have some buffer).

All of this is (you guessed it) UNTESTED and will take a lot of ironing
out tomorrow. But now it's 1 AM again and I need to sleeeeeep