Bugfix: don't error when crossing off-path tiles
Catch a mis-re-raised ValueError (re-raised as IndexError in advance()
because that makes a lot more sense to me) that ToolheadTrajectory
throws when a tile is not on the path. This was causing issues right
at startup because the toolhead wasn't on the path yet. This should be
fine -- we're using absolute-coordinate moves so the first motion
segment will just cross any tiles it needs to to get to the start
of the trajectory.