Clean up logs; enqueue multiple tiles when appropriate; test toolhead
Thus far I'm seeing us get motion reports frequently enough that we
catch each tile, but on reflection this is likely to not be the case
pretty frequently when we make fast moves or there's any kind of
hiccup in the websocket. That's fine, we'll just keep a deep enough
buffer (currently set to 7, that might be longer than we need) that
we can chew it up a little and then enqueue to depth again. Add a
"number of tiles advanced" return value and use it to keep buffer depth.
(Will this ever get off? I'm not sure hm)

Also clean up some logging so that the resulting log is nice and pretty.
A pretty log is important! Key source of visibility!

This is now working pretty well! I'm also checking in the script
I've been using to test just in case that's useful.