Fix re-routing in flight (and some debug logs)
I'd broken Zoda's go_to_coords() debugging in converting from print
statements, my bad! fix that.

Also rename "self.paused" on Toolhead to be more clear what it's
actually doing. I really don't like that mechanism, would love to
get rid of the statefulness there and in "self.on_path" also. Feels
like workarounds. I'll see if I can simplify that going forward.

Anyway, fix that logic too. Now we should be able to re-route in
flight without issue.

Another issue here was that go_to_coords() is (I think) including
the starting tile in its returned path, which makes sense but was
causing bugs when it was passed as-is to extend() and we ended up
with the same tile twice (the mouse was getting stuck on zero-length
movements I think). Add a little escape hatch to extend() to catch
that (is that the most elegant place to do it? IDK)

FINALLY in manual_control, don't clear() the trajectory unless
we've actually gotten a valid new move, so that clicking on an
obstacle does nothing (previously would clear the trajectory and
fail to extend it).

TODO: Actually track the buffer depth of submitted gcode segments
properly so that we can restore to MOVE_BUFFER_DEPTH_TILES when
new tiles are submitted to the trajectory.