Port ArcSquiggles to a MotionStyle
I think it'd be fun to use ArcSquiggles for regular mousin' around
and to then use a faster motion to run away. This commit ports ArcSquiggles
to a MotionStyle. Note that this is going to cause some stuttering
because MotionStyles are fed paths a few tiles at a time, rather than
getting the full path, and ArcSquiggles makes heavy use of lookahead
for cornering. I think this means that we should mumble mumble lookahead
but we'll figure that out later, for now it'll just stutter and move
weirdly and that's what it's gonna be.